Fix caching bug in server

+++ api/Engraved.Core/Source/Application/Commands/Measurements/Delete/DeleteMeasurementCommandExecutor.cs
@@ -1,4 +1,6 @@
 ﻿using Engraved.Core.Application.Persistence;
+using Engraved.Core.Domain.Measurements;
+using Engraved.Core.Domain.Metrics;
 
 namespace Engraved.Core.Application.Commands.Measurements.Delete;
 
@@ -13,8 +15,16 @@ public class DeleteMeasurementCommandExecutor : ICommandExecutor
 
   public async Task<CommandResult> Execute(IRepository repository, IDateService dateService)
   {
+    IMeasurement? measurement = await repository.GetMeasurement(_command.Id);
+    if (measurement == null)
+    {
+      return new CommandResult();
+    }
+
     await repository.DeleteMeasurement(_command.Id);
 
-    return new CommandResult { EntityId = _command.Id };
+    IMetric metric = (await repository.GetMetric(measurement.MetricId))!;
+
+    return new CommandResult( _command.Id, metric.Permissions.GetUserIdsWithAccess());
   }
 }
